+++ core-service/core-api/src/main/java/ru/home/aglar/market/core/api/OrderItemDto.java
@@ -2,22 +2,24 @@ package ru.home.aglar.market.core.api;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 
+import java.math.BigDecimal;
+
 public class OrderItemDto {
     @Schema(description = "Id of product", requiredMode = Schema.RequiredMode.REQUIRED)
     private Long productId;
     @Schema(description = "Title", requiredMode = Schema.RequiredMode.REQUIRED)
     private String title;
     @Schema(description = "Price of one product", requiredMode = Schema.RequiredMode.REQUIRED)
-    private Integer price;
+    private BigDecimal price;
     @Schema(description = "Quantity of products", requiredMode = Schema.RequiredMode.REQUIRED)
     private Integer quantity;
     @Schema(description = "Total price all products in this item", requiredMode = Schema.RequiredMode.REQUIRED)
-    private Integer totalPrice;
+    private BigDecimal totalPrice;
 
     public OrderItemDto() {
     }
 
-    public OrderItemDto(Long productId, String title, Integer price, Integer quantity, Integer totalPrice) {
+    public OrderItemDto(Long productId, String title, BigDecimal price, Integer quantity, BigDecimal totalPrice) {
         this.productId = productId;
         this.title = title;
         this.price = price;
@@ -41,11 +43,11 @@ public class OrderItemDto {
         this.title = title;
     }
 
-    public Integer getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(Integer price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
 
@@ -57,11 +59,11 @@ public class OrderItemDto {
         this.quantity = quantity;
     }
 
-    public Integer getTotalPrice() {
+    public BigDecimal getTotalPrice() {
         return totalPrice;
     }
 
-    public void setTotalPrice(Integer totalPrice) {
+    public void setTotalPrice(BigDecimal totalPrice) {
         this.totalPrice = totalPrice;
     }
 }
